Include credit % in unrealised profitability (#1152)
- Make sure accrued interest if reflected in the strategy profitability %

+++ tradeexecutor/visual/equity_curve.py
@@ -523,7 +523,7 @@ def calculate_compounding_unrealised_trading_profitability(
     profit_data.sort(key=lambda t: t[0])
 
     index, profit = list(zip(*profit_data))
-    returns = pd.Series(data=profit, index=pd.DatetimeIndex(index))
+    returns = pd.Series(data=profit, index=pd.DatetimeIndex(index), dtype="float64")
 
     if freq:
